+++ src/Seomatic.php
@@ -19,6 +19,7 @@ use nystudio107\seomatic\gql\interfaces\SeomaticInterface;
 use nystudio107\seomatic\gql\resolvers\SeomaticResolver;
 use nystudio107\seomatic\gql\queries\SeomaticQuery;
 use nystudio107\seomatic\helpers\Environment as EnvironmentHelper;
+use nystudio107\seomatic\helpers\Gql as GqlHelper;
 use nystudio107\seomatic\helpers\MetaValue as MetaValueHelper;
 use nystudio107\seomatic\integrations\feedme\SeoSettings as SeoSettingsFeedMe;
 use nystudio107\seomatic\listeners\GetCraftQLSchema;
@@ -55,6 +56,7 @@ use craft\events\DefineGqlTypeFieldsEvent;
 use craft\events\RegisterCacheOptionsEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterGqlQueriesEvent;
+use craft\events\RegisterGqlSchemaComponentsEvent;
 use craft\events\RegisterGqlTypesEvent;
 use craft\events\RegisterPreviewTargetsEvent;
 use craft\events\RegisterUrlRulesEvent;
@@ -267,7 +269,7 @@ class Seomatic extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '3.0.9';
+    public $schemaVersion = '3.0.10';
 
     /**
      * @var bool
@@ -677,6 +679,19 @@ class Seomatic extends Plugin
                     }
                 }
             );
+            // Handler: Gql::EVENT_REGISTER_SCHEMA_COMPONENTS
+            Event::on(
+                Gql::class,
+                Gql::EVENT_REGISTER_GQL_SCHEMA_COMPONENTS,
+                function (RegisterGqlSchemaComponentsEvent $event) {
+                    Craft::debug(
+                        'Gql::EVENT_REGISTER_GQL_SCHEMA_COMPONENTS',
+                        __METHOD__
+                    );
+                    $label = Craft::t('seomatic', 'Seomatic');
+                    $event->queries[$label]['seomatic.all:read'] = ['label' => Craft::t('seomatic', 'Query Seomatic data')];
+                }
+            );
         }
         // Add support for querying for SEOmatic metadata inside of element queries
         if (self::$craft34) {
@@ -690,14 +705,17 @@ class Seomatic extends Plugin
                         'TypeManager::EVENT_DEFINE_GQL_TYPE_FIELDS',
                         __METHOD__
                     );
-                    // Make Seomatic tags available to all entries.
-                    $event->fields['seomatic'] = [
-                        'name' => 'seomatic',
-                        'type' => SeomaticInterface::getType(),
-                        'args' => SeomaticArguments::getArguments(),
-                        'resolve' => SeomaticResolver::class . '::resolve',
-                        'description' => Craft::t('seomatic', 'This query is used to query for SEOmatic meta data.')
-                    ];
+
+                    if (GqlHelper::canQuerySeo()) {
+                        // Make Seomatic tags available to all entries.
+                        $event->fields['seomatic'] = [
+                            'name' => 'seomatic',
+                            'type' => SeomaticInterface::getType(),
+                            'args' => SeomaticArguments::getArguments(),
+                            'resolve' => SeomaticResolver::class . '::resolve',
+                            'description' => Craft::t('seomatic', 'This query is used to query for SEOmatic meta data.')
+                        ];
+                    }
                 }
             });
         }