+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # SEOmatic Changelog
 
+## 3.3.43 - 2021.06.08
+### Added
+* Add GraphQL schema components for SEOmatic.
+
+### Changed
+* Switched documentation system to VitePress
+
+### Fixed
+* Additional Sitemap URLs will now appear only in the Site's sitemap index where they are defined (https://github.com/nystudio107/craft-seomatic/issues/897)
+
 ## 3.3.42 - 2021.05.18
 ### Changed
 * Strip the query params from next/prev pagination meta tags if we're not using query-string based pagination (https://github.com/nystudio107/craft-seomatic/issues/896)